Add function of defending navi page when making user's Graph

+++ RabbitBackFlask/rabbit/fox.py
@@ -49,7 +49,7 @@ import way
 
 def uploadGoogle(userId_val):
     import socket
-    socket.setdefaulttimeout(60*10) # Graph upload time: 5 minutes limit
+    socket.setdefaulttimeout(60*15) # Graph upload time: 15 minutes limit
     
     folder_id = '1Pfj_Qaf8HBHqBQ9hNcIdPCf_dPhmGspY'
 
@@ -151,18 +151,28 @@ def calcCrime(crime_val, userId_val):
 
 @app.route('/api/navi/', methods=['POST'])
 def apiNavi():
-    orig = request.json['orig']
-    dest = request.json['dest']
     id = request.json['id']
-    
-    p1 = Point(orig['longitude'], orig['latitude'])
-    p2 = Point(dest['longitude'], dest['latitude'])
 
-    if (p1.within(poly) and p2.within(poly)):
-        return jsonify(way.wayNine(orig, dest, str(id))) # In this func, return 9 + 9 + 9 + 9 waypoints
-    else: # Error 400
-        return 'Error, Please enter correct Chicago Coordinates', status.HTTP_400_BAD_REQUEST
+    cur = mysql.connection.cursor()
+    cur.execute("SELECT nowStatus FROM user WHERE id = '%s'" % (id))
+    nowStatus = cur.fetchone()
+    nowStatus = int(nowStatus['nowStatus'])
+    cur.close()
+
+    if nowStatus == 0:
+        orig = request.json['orig']
+        dest = request.json['dest']
         
+        p1 = Point(orig['longitude'], orig['latitude'])
+        p2 = Point(dest['longitude'], dest['latitude'])
+
+        if (p1.within(poly) and p2.within(poly)):
+            return jsonify(way.wayNine(orig, dest, str(id))) # In this func, return 9 + 9 + 9 + 9 waypoints
+        else: # Error 400
+            return 'Error, Please enter correct Chicago Coordinates', status.HTTP_400_BAD_REQUEST
+    else: # nowStatus == 1
+        return 'Error, Please wait until revising your graph.', status.HTTP_400_BAD_REQUEST
+
 
 @app.route('/api/signup/calculate/', methods=['POST'])  # not exist? then save. After Sign-Up process, user saved. Then this request can be used.
 def apiCalc():
@@ -192,6 +202,8 @@ def apiCalc():
         fileId = calcCrime(crime, userId)
         cur.execute("INSERT INTO fileinfo (user_id, file_id) VALUES ('%s', '%s')" % (userId, fileId))
 
+    cur.execute("UPDATE user SET nowStatus = 0 WHERE id = '%s'" % (userId)) # nowStatus update (done making user's graph)
+    
     mysql.connection.commit()
     cur.close()
 
@@ -204,26 +216,34 @@ def apiDBUpdate():
     crime = str(request.json['crime'])
     userId = str(request.json['id'])
 
-    fileId = calcCrime(crime, userId)
+    if len(crime) == 0:
+        return 'Error, Please input at least one crime option', status.HTTP_400_BAD_REQUEST
+    else:
+        print('I am here')
+        cur = mysql.connection.cursor()
+        cur.execute("UPDATE user SET nowStatus = 1 WHERE id = '%s'" % (userId)) # nowStatus update (making user's graph)
+        mysql.connection.commit()
 
-    cur = mysql.connection.cursor()
-    
-    # Delete previous User Graph file in Google Drive
-    cur.execute("SELECT file_id FROM fileinfo WHERE user_id = '%s'" % (userId))
-    pre_file_id = cur.fetchone()
-    pre_file_id = str(pre_file_id['file_id'])
-    service.files().delete(fileId=pre_file_id).execute()
-    
-    # Update new User Graph file id
-    cur.execute("UPDATE fileinfo SET file_id = '%s' WHERE user_id = '%s'" % (fileId, userId))
+        fileId = calcCrime(crime, userId)
+        
+        # Delete previous User Graph file in Google Drive
+        cur.execute("SELECT file_id FROM fileinfo WHERE user_id = '%s'" % (userId))
+        pre_file_id = cur.fetchone()
+        pre_file_id = str(pre_file_id['file_id'])
+        service.files().delete(fileId=pre_file_id).execute()
+        
+        # Update new User Graph file id
+        cur.execute("UPDATE fileinfo SET file_id = '%s' WHERE user_id = '%s'" % (fileId, userId))
 
-    # Change 'user' table crime value
-    cur.execute("UPDATE user SET crime = '%s' WHERE id = '%s'" % (crime, userId))
-    
-    mysql.connection.commit()
-    cur.close()
+        # Change 'user' table crime value
+        cur.execute("UPDATE user SET crime = '%s' WHERE id = '%s'" % (crime, userId))
+        
+        cur.execute("UPDATE user SET nowStatus = 0 WHERE id = '%s'" % (userId))
 
-    return "Done UPDATE crime DB"
+        mysql.connection.commit()
+        cur.close()
+
+        return "Done UPDATE crime DB"
 
 
 if __name__=='__main__':
